Integrate Joi validation by promoting Quiz state to demo app wrapper
The app wrapper was necessary because validating state on change while adding error info to state leads to infinite loops.

The quiz validator packaged with kasra-validators is currently unsuitable for use here, because it is designed to validate an article containing a quiz, not a quiz. The validator should be split up when time permits.